Se separo otherInfoWindow en una primera version de la vista y el modelo
